fix: Avoid logging file contents during uploads (#201)
If the request is multipart, don't log the body contents. This
doesn't prevent logging the file in absolutely all cases (e.g. a
PUT with the contents as the body), but it will omit the messages
as long as the upload comes from the WunderGraph SDK. Added notes
in both places to make sure they're kept in sync.