fix: avoid logging file contents during uploads (#201) (#203)
* fix: Avoid logging file contents during uploads (#201)

If the request is multipart, don't log the body contents. This
doesn't prevent logging the file in absolutely all cases (e.g. a
PUT with the contents as the body), but it will omit the messages
as long as the upload comes from the WunderGraph SDK. Added notes
in both places to make sure they're kept in sync.

* fix: Split logger middleware into a free function, add test

* fix: Remove unneeded comment in uploadFiles()

* fix: Document logRequestMiddleware()

* Make the test for multipart requests a bit more specific

Test for multipart/form-data instead of multipart/

* refactor: Change TestLogMiddleware_Debug() to use subtests and asserts

Co-authored-by: Dustin Deus <deusdustin@gmail.com>