feat: ensure wunderctl always call the same copy when calling itself (#300)
* feat: ensure wunderctl always call the same copy when calling itself

Introduce a WUNDERCTL_BINARY_PATH environment variable. This can
be set either by the user or either by wunderctl itself. On startup,
wunderctl checks if the variable is unset and if it is, then it
sets it to its own path. The TS side now checks for this new variable
and when it's non-empty, it overrides the path to the wunderctl binary.

* refactor: use os.LookupEnv instead of iterating the environment

* refactor: pass WUNDERCTL_BINARY_PATH to each subprocess individually

* refactor: retrieve wunderctl path when spawning children

Avoids using a global variable