feat: add support for parallel testing.Server instances
Use random ports for the node and the hooks server, so we can run
several instances at the same time. There's a small chance of a race
condition since we're picking the port in the TS side and using it in
the Go side, but it shouldn't be a problem in practical terms.