fix: print compiler error messages during TS operation response schema generation
Previously, we were silencing error messages from typescript-json-schema because
we have some fallback logic to handle some errors. This is a problem when
we get an error that cannot be recovered, because we were not printing what the
error actually was. Now we collect the error messages, apply deduplication and
we print them using console.error() before returning the failure.