test: fix changed files detection for external PRs (#875)
Use a simpler approach to determine the refs that define the interval
of the changes contained in either the PR of the push.
For pull requests, use github.event.pull_request.base.sha and for
pushes just compare HEAD against its previous commit. This should
for the both use cases we support with external repos.