fix: wait for the hooks server to start up before listening on the node
Hooks server and node were started in parallel. This causes a problem
when the node is ready and starts listening before the hooks server has
had time to do its own setup and opening its port, because operations that
call into the hooks server will fail inmmediately due to the connection
to the hooks server being refused.

To fix this, wait for the hooks server to be ready (by connecting to
its port) during startup for wunderctl up and wunderctl start (iff the
hooks server is also started by us).