feat: use Prometheus to track and expose operation and API request counts (#991)
Add support for exposing metrics using Prometheus, which can be optionally disabled at runtime. The following metrics are exposed:

For operations:
- Incoming HTTP request duration
- Incoming HTTP request count (with status code)
- Incoming HTTP request size
- Outgoing HTTP response size

For upstreams:
- Number of requests sent per host (with method and status code)
- Duration of requests sent per host

Prometheus metrics are on by default, listening on http://<host>:8881/metrics. They can be disabled when calling
configureWunderGraphApplication() and they're also controlled by environment variables.

Additionally, add a guide about monitoring and observability with a section about Prometheus, to be extended
later when we add OTEL support.