feat: support different types with the same when generating OAS
Previously we were just detecting if two types with the same name were
different and failing in that case, under the assumption that the user
could always renamed the types. Unfortunately, this turned out to not be
possible when the types are generated by some upstream APIs.

To support this use case, rename the new type by appending a number and
rename also all its references (within the schema component and its children).
This way the names won't collide anymore.