feat: add support for server ping during subscription idle times
Some provides or middleboxes aggressively close TCP connections that
look inactive. This means a long running subscription might be end
disconnected if no new data arrives for some time. To counteract this,
introduce a ping mechanism in the server that ensures that the line is
not idle for longer than the requested interval by sending an empty
payload if no actual data has been sent.

Also, refactor all the SSE / jsonpatch support to implement it in a single
place.

Additionally, fix a bug which incorrectly shared a variable in the client
when returning subscription results to the caller.