feat: propagate error messages from functions/hooks to API responses (#1248)
When a function or a hook throws an error, propagate it through the stack
until it's output by the API as an error. This allows clients to retrieve
the very same error message that was produced in the backend.

If the error is not one of the types we control, just log it to the terminal's
output without returning it through the API, since that could expose internal
implementation details.