feat: upgrade @graph-mesh and @omnigraph dependencies, update patches too (#1331)
This drops patches that have been already merged upstream and adds a patch
for https://github.com/ardatan/graphql-mesh/pull/6231, which is not yet released

Format for the HTTP error messages has changed slightly in new version of graphql-mesh. This shouldn't
affect clients of our library because we don't expose these types.